fixed cleaning script and matched it with .inx config
Changed Provider name in .inx file to match clean script.

+++ windows/driver/package/clean64.ps1
@@ -1,14 +1,19 @@
 .\devcon64.exe remove root\wudfvhidmini
-$infs = (pnputil -e) | Out-String
-$resultlist = New-Object System.Collections.Specialized.StringCollection
+$pnputil = (pnputil -e ) | Out-String
+$inflist = New-Object System.Collections.Specialized.StringCollection
 $regex = [regex] '(?m)^.*(oem\d+\.inf).*\r\n.*f3flight'
-$match = $regex.Match($subject)
+$match = $regex.Match($pnputil)
 while ($match.Success) {
-	$resultlist.Add($match.Groups[1].Value) | Out-Null
+	$inflist.Add($match.Groups[1].Value) | Out-Null
 	$match = $match.NextMatch()
 }
-foreach ($inf in $infs)
+
+foreach ($inf in $inflist)
 {
+    echo $inf
 	pnputil -d $inf
 }
-Remove-Item C:\Windows\System32\drivers\UMDF\WUDFvhidmini.dll
+
+if (Test-Path "C:\Windows\System32\drivers\UMDF\WUDFvhidmini.dll") {
+    Remove-Item C:\Windows\System32\drivers\UMDF\WUDFvhidmini.dll
+}
